requirements.txt
aiogram==3.12.0
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
pydantic==2.8.2  # важно: aiogram 3.12 несовместим с 2.9+
python-dotenv==1.0.1
redis==5.0.8
loguru==0.7.2

# Этапы 2+ (пока закомментируем, чтобы не мешали сборке)
APScheduler==3.10.4
pytz==2024.1

# APScheduler==3.10.4
# Jinja2==3.1.4
# WeasyPrint==62.3
pytest
pytest-asyncio
openai>=1.6.0
httpx>=0.27
